Revert state_space to float32; it is a numerical-correctness fix, not style
A prior task widened both observation_space and state_space to float64 in
cartpole.py and quadrotor.py. observation_space is correctly float64 -- the
envs genuinely emit float64 and check_env requires the declaration to match
-- but state_space must stay float32.

state_space.low/high are read by out-of-bounds logic (quadrotor.py's
_get_done) and by DefaultConstraint (safe_control_gym/envs/constraints.py),
and are written into by the trajectory generators
(generate_quadrotor_3d_trajectories.py, generate_quadrotor_2d_visualization.py,
calibrate_quadrotor_{2d,3d}_bounds.py). Those writes previously rounded to
float32; storing them at float64 instead keeps the exact value, shifting
thresholds by up to 1.9e-7 and flipping at least one out_of_bounds decision.
Since this repo's binding constraint for the gymnasium migration is that
nothing changes numerically, that drift is a regression.

check_env never inspects state_space, so gymnasium conformance
(test_gymnasium_conformance.py, 4/4) is unaffected by the revert.

+++ safe_control_gym/envs/gym_control/cartpole.py
@@ -497,7 +497,7 @@ class CartPole(BenchmarkEnv):
                               self.x_dot_threshold,
                               self.theta_threshold_radians * 2,
                               self.theta_dot_threshold])
-        self.state_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float64)
+        self.state_space = spaces.Box(low=-obs_bound, high=obs_bound, dtype=np.float32)
 
         # Concatenate goal info for RL
         if self.COST == Cost.RL_REWARD and self.TASK == Task.TRAJ_TRACKING and self.obs_goal_horizon > 0:
